+++ src/main/java/com/fbr1/mesasutnfrro/model/data/ExamenRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
-public interface ExamenRepository extends CrudRepository<Examen, Integer>{
+public interface ExamenRepository extends CrudRepository<Examen, Long>{
 }
